Update gen scripts
