+++ {{cookiecutter.collection_name}}/docs/gen_examples_catalog.py
@@ -1,6 +1,5 @@
 """
-Copies README.md to index.md. Also discovers all blocks and
-generates a list of them in the docs under the Blocks Catalog heading.
+Locates all the examples in the Collection and puts them in a single page.
 """
 
 import re
@@ -8,120 +7,26 @@ from collections import defaultdict
 from inspect import getmembers, isclass, isfunction, ismodule
 from pathlib import Path
 from textwrap import dedent
-from typing import Any, Set
+from types import ModuleType
+from typing import Callable, Set, Union
 
 import mkdocs_gen_files
 from griffe.dataclasses import Docstring
 from griffe.docstrings.dataclasses import DocstringSectionKind
 from griffe.docstrings.parsers import Parser, parse
-from prefect.blocks.core import Block
 from prefect.logging.loggers import disable_logger
-from prefect.utilities.dispatch import get_registry_for_type
 from prefect.utilities.importtools import to_qualified_name
 
 import {{ cookiecutter.collection_slug }}
 
 COLLECTION_SLUG = "{{ cookiecutter.collection_slug }}"
 
-# Home page
 
-
-readme_path = Path("README.md")
-docs_index_path = Path("index.md")
-
-with open(readme_path, "r") as readme:
-    with mkdocs_gen_files.open(docs_index_path, "w") as generated_file:
-        for line in readme:
-            if line.startswith("Visit the full docs [here]("):
-                continue  # prevent linking to itself
-            generated_file.write(line)
-
-    mkdocs_gen_files.set_edit_path(Path(docs_index_path), readme_path)
-
-# Blocks Catalog page
-
-
-def find_module_blocks():
-    blocks = get_registry_for_type(Block)
-    collection_blocks = [
-        block
-        for block in blocks.values()
-        if to_qualified_name(block).startswith(COLLECTION_SLUG)
-    ]
-    module_blocks = {}
-    for block in collection_blocks:
-        block_name = block.__name__
-        module_nesting = tuple(to_qualified_name(block).split(".")[1:-1])
-        if module_nesting not in module_blocks:
-            module_blocks[module_nesting] = []
-        module_blocks[module_nesting].append(block_name)
-    return module_blocks
-
-
-def insert_blocks_catalog(generated_file):
-    module_blocks = find_module_blocks()
-    if len(module_blocks) == 0:
-        return
-    generated_file.write(
-        dedent(
-            f"""
-            Below is a list of Blocks available for registration in
-            `"{{ cookiecutter.collection_name }}"`.
-
-            To register blocks in this module to
-            [view and edit them](https://orion-docs.prefect.io/ui/blocks/)
-            on Prefect Cloud:
-            ```bash
-            prefect block register -m {COLLECTION_SLUG}
-            ```
-            """
-        )
-    )
-    generated_file.write(
-        "Note, to use the `load` method on Blocks, you must already have a block document "  # noqa
-        "[saved through code](https://orion-docs.prefect.io/concepts/blocks/#saving-blocks) "  # noqa
-        "or [saved through the UI](https://orion-docs.prefect.io/ui/blocks/).\n"
-    )
-    for module_nesting, block_names in module_blocks.items():
-        module_path = " ".join(module_nesting)
-        module_title = module_path.replace("_", " ").title()
-        generated_file.write(
-            f"## [{module_title} Module][{COLLECTION_SLUG}.{module_path}]\n"
-        )
-        for block_name in block_names:
-            generated_file.write(
-                f"[{block_name}][{COLLECTION_SLUG}.{module_path}.{block_name}]\n"
-            )
-            generated_file.write(
-                dedent(
-                    f"""
-                    To load the {block_name}:
-                    ```python
-                    from prefect import flow
-                    from {COLLECTION_SLUG}.{module_path} import {block_name}
-
-                    @flow
-                    def my_flow():
-                        my_block = {block_name}.load("MY_BLOCK_NAME")
-
-                    my_flow()
-                    ```
-                    """
-                )
-            )
-
-
-blocks_catalog_path = Path("blocks_catalog.md")
-with mkdocs_gen_files.open(blocks_catalog_path, "w") as generated_file:
-    insert_blocks_catalog(generated_file)
-
-# Examples Catalog page
-
-
-def skip_parsing(name, obj, module_nesting):
+def skip_parsing(name: str, obj: Union[ModuleType, Callable], module_nesting: str):
     """
-    Skips parsing the object if it's a private method or if it's not in the
-    right module.
+    Skips parsing the object if it's a private object or if it's not in the
+    module nesting, preventing imports from other libraries from being added to the
+    examples catalog.
     """
     try:
         wrong_module = not to_qualified_name(obj).startswith(module_nesting)
@@ -130,14 +35,14 @@ def skip_parsing(name, obj, module_nesting):
     return obj.__doc__ is None or name.startswith("_") or wrong_module
 
 
-def skip_code_example(code_example: str) -> bool:
+def skip_block_load_code_example(code_example: str) -> bool:
     """
     Skips the code example if it's just showing how to load a Block.
     """
     return re.search(r'\.load\("BLOCK_NAME"\)\s*$', code_example.rstrip("`"))
 
 
-def get_code_examples(obj: Any) -> Set[str]:
+def get_code_examples(obj: Union[ModuleType, Callable]) -> Set[str]:
     """
     Gathers all the code examples within an object.
     """
@@ -152,13 +57,13 @@ def get_code_examples(obj: Any) -> Set[str]:
             code_example = "\n".join(
                 (part[1] for part in section.as_dict().get("value", []))
             )
-            if not skip_code_example(code_example):
+            if not skip_block_load_code_example(code_example):
                 code_examples.add(code_example)
         if section.kind == DocstringSectionKind.admonition:
             value = section.as_dict().get("value", {})
             if value.get("annotation") == "example":
                 code_example = value.get("description")
-                if not skip_code_example(code_example):
+                if not skip_block_load_code_example(code_example):
                     code_examples.add(code_example)
 
     return code_examples
@@ -188,6 +93,7 @@ for module_name, module_obj in getmembers({{ cookiecutter.collection_slug }}, is
         if skip_parsing(function_name, function_obj, module_nesting):
             continue
         code_examples_grouping[module_name] |= get_code_examples(function_obj)
+
 
 examples_catalog_path = Path("examples_catalog.md")
 with mkdocs_gen_files.open(examples_catalog_path, "w") as generated_file:
